splitup sbe_proc.py
